consistent pricing table across onboarding and settings
